fix: refactor modal and save searches
