+++ packages/frontend/src/components/Backdrop/index.ts
@@ -1,2 +1,1 @@
 export { Backdrop } from './Backdrop.tsx';
-export { ModalWithBackdrop } from './Modal.tsx';
