+++ packages/frontend/src/i18n/resources/en.json
@@ -75,8 +75,10 @@
   "route.inbox": "Inbox",
   "route.sent": "Sent",
   "savedSearches.change_name": "Change name",
-  "savedSearches.confirmDelete": "Are you sure you want to delete the saved search??",
-  "savedSearches.confirm_delete_title": "Confirm delete",
+  "savedSearches.confirm_delete_body": "Are you sure you want to delete the saved search??",
+  "savedSearches.confirm_delete_title": "Delete Saved Search",
+  "savedSearches.buttons.cancel": "Cancel",
+  "savedSearches.buttons.confirm": "Yes, continue",
   "savedSearches.delete_search": "Delete saved search",
   "savedSearches.deleted_failed": "Deleting search failed",
   "savedSearches.deleted_success": "Search deleted",
