+++ packages/frontend/src/i18n/resources/nb.json
@@ -75,8 +75,10 @@
   "route.inbox": "Innboks",
   "route.sent": "Sendt",
   "savedSearches.change_name": "Rediger søk",
-  "savedSearches.confirmDelete": "Er du sikker på at du vil slette søket?",
-  "savedSearches.confirm_delete_title": "Bekreft sletting",
+  "savedSearches.confirm_delete_body": "Er du sikker på at du vil slette søket?",
+  "savedSearches.buttons.cancel": "Avbryt",
+  "savedSearches.buttons.confirm": "Ja, forsett",
+  "savedSearches.confirm_delete_title": "Slett lagret søk",
   "savedSearches.delete_search": "Slett",
   "savedSearches.deleted_failed": "Sletting av søk feilet",
   "savedSearches.deleted_success": "Søk slettet",
