+++ packages/frontend/src/pages/SavedSearches/EditSavedSearchDialog/editSavedSearchDialog.module.css
@@ -1,0 +1,87 @@
+.searchDetails {
+  text-align: center;
+  font-size: 0.875rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 1.3;
+  align-self: center;
+}
+
+.searchString {
+  border-bottom: 1px solid #000;
+}
+
+.filterElement {
+  border: 1px solid #000;
+  border-radius: 0.125rem;
+  padding: 0.25rem 0.625rem;
+}
+
+.lastUpdated {
+  color: var(--black-50, rgba(0, 0, 0, 0.5));
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 1.5;
+  margin: 20px 0 0;
+}
+
+.icon {
+  width: 24px;
+  height: 24px;
+}
+
+.searchFormBody {
+  margin-bottom: 16px;
+
+  input {
+    box-sizing: border-box;
+  }
+}
+
+.searchFormBody label {
+  color: var(--black-100, #000);
+  font-size: 0.875rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 0.8;
+  padding-bottom: 1rem;
+}
+
+.searchFormBody input[type="text"] {
+  width: 100%;
+  padding: 8px;
+  margin-bottom: 16px;
+  margin-top: 16px;
+  border-radius: 4px;
+  border: var(--border_width-default, 1px) solid #3381c8;
+  background: #fff;
+  color: #000;
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
+
+.searchFormFooter {
+  display: grid;
+  grid-template-columns: auto auto auto;
+  align-items: center;
+}
+
+.buttons {
+  display: flex;
+  column-gap: 1rem;
+  margin-right: 1rem;
+}
+
+.updateTime {
+  color: var(--black-50, rgba(0, 0, 0, 0.5));
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 1.5;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
