+++ packages/frontend/src/pages/SavedSearches/SavedSearchesActions/savedSearchesActions.module.css
@@ -1,0 +1,50 @@
+.renderButtons {
+  display: flex;
+  align-items: center;
+}
+
+.linkButton {
+  border: none;
+  background-color: #fff;
+  color: var(--black-50, rgba(0, 0, 0, 0.5));
+  cursor: pointer;
+}
+
+.linkButton:hover {
+  background-color: #fff;
+  color: var(--black-50, rgba(0, 0, 0, 0.5));
+}
+
+.icon {
+  width: 24px;
+  height: 24px;
+}
+
+.dropdownContent {
+  border-radius: 0;
+
+  li {
+    margin: 0;
+  }
+
+  button {
+    border-radius: 2px;
+    padding: 0;
+    height: auto;
+  }
+}
+
+.dropdownEditSearch {
+  width: 100%;
+  display: grid;
+  grid-template-columns: min-content auto min-content;
+  align-items: start;
+  text-align: left;
+  gap: 6px;
+  padding: 6px;
+
+  & > svg,
+  & > span {
+    padding: 4px;
+  }
+}
