+++ packages/frontend/src/pages/SavedSearches/SearchFilterTag/searchFilterTag.module.css
@@ -1,0 +1,5 @@
+.searchFilterTag {
+  border: 1px solid #000;
+  border-radius: 0.125rem;
+  padding: 0.25rem 0.625rem;
+}
