+++ packages/frontend/src/pages/SavedSearches/savedSearchesPage.module.css
@@ -1,0 +1,131 @@
+.savedSearchesWrapper {
+  background: var(--Company-300);
+  padding: 1rem;
+}
+
+.savedSearchesList {
+  background: #fff;
+  border-radius: 4px;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+}
+
+.savedSearchItem {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 0 1rem;
+}
+
+.title {
+  color: #000;
+  font-weight: 500;
+  font-size: 1.5rem;
+  line-height: 1.25;
+  margin: 1rem 0;
+}
+
+.searchDetails {
+  text-align: center;
+  font-size: 0.875rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 1.3;
+  align-self: center;
+}
+
+.searchString {
+  border-bottom: 1px solid #000;
+}
+
+.lastUpdated {
+  color: var(--black-50, rgba(0, 0, 0, 0.5));
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 1.5;
+  margin: 20px 0 0;
+}
+
+.icon {
+  width: 24px;
+  height: 24px;
+}
+
+.searchFormHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 16px;
+}
+
+.searchTerm {
+  font-weight: bold;
+}
+
+.searchStatus {
+  background-color: #f0f0f0;
+  padding: 4px 8px;
+  border-radius: 4px;
+}
+
+.closeButton {
+  background: none;
+  border: none;
+  font-size: 1rem;
+  cursor: pointer;
+}
+
+.searchFormBody {
+  margin-bottom: 16px;
+
+  input {
+    box-sizing: border-box;
+  }
+}
+
+.searchFormBody label {
+  color: var(--black-100, #000);
+  font-size: 0.875rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 0.8;
+  padding-bottom: 1rem;
+}
+
+.searchFormBody input[type="text"] {
+  width: 100%;
+  padding: 8px;
+  margin-bottom: 16px;
+  margin-top: 16px;
+  border-radius: 4px;
+  border: var(--border_width-default, 1px) solid #3381c8;
+  background: #fff;
+  color: #000;
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
+
+.searchFormFooter {
+  display: grid;
+  grid-template-columns: auto auto auto;
+  align-items: center;
+}
+
+.saveButton,
+.deleteButton {
+  margin-right: 1rem;
+}
+
+.updateTime {
+  color: var(--black-50, rgba(0, 0, 0, 0.5));
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 1.5;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
